add prevent delete in methos eliminarUsuario

+++ src/main/resources/static/js/usuariosv5.js
@@ -34,11 +34,15 @@ async function cargarUsuarios(){
 };
 
 async function eliminarUsuario(id){
-          const request = await fetch('api/usuarios/' + id, {
-                  method: 'DELETE',
-                  headers: {
-                    'Accept': 'application/json',
-                    'Content-Type': 'application/json'
-                  },
-                });
-      }
+    if(!confirm('¿Desea Eliminar este usuario?')){
+     return
+    }
+     const request = await fetch('api/usuarios/' + id, {
+                      method: 'DELETE',
+                      headers: {
+                        'Accept': 'application/json',
+                        'Content-Type': 'application/json'
+                      },
+                    });
+     location.reload()
+}
